Relocate 3 presenters
PayMeansOfProductionPresenter, ListPlansPresenter  and HidePlanPresenter
