+++ arbeitszeit_web/presenters/hide_plan_presenter.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from arbeitszeit.use_cases.hide_plan import HidePlanResponse
 from arbeitszeit_web.translator import Translator
 
-from .notification import Notifier
+from ..notification import Notifier
 
 
 @dataclass
