+++ tests/presenters/test_list_plans_presenter.py
@@ -2,7 +2,10 @@ from unittest import TestCase
 from uuid import uuid4
 
 from arbeitszeit.use_cases.list_active_plans_of_company import ListedPlan
-from arbeitszeit_web.list_plans import ListPlansPresenter, ListPlansResponse
+from arbeitszeit_web.presenters.list_plans_presenter import (
+    ListPlansPresenter,
+    ListPlansResponse,
+)
 
 from .dependency_injection import get_dependency_injector
